+++ test/python/additive.py
@@ -47,3 +47,31 @@ def test_trace_unequal_arity():
 
     f = Function(inside, (int, int), (int, int, int))
     assert f.trace()(7) == (7, 1)
+
+
+def test_strategy():
+    from hypothesis import find
+    from discopy.python.additive import Function
+
+    find(Function.strategy(), lambda f: len(f.dom) > len(f.cod) >= 1)
+
+
+def test_Types_matmul():
+    from discopy.python.additive import Ty
+
+    assert (int,) @ Ty((str,)) == Ty((int, str)) == Ty((int,)) @ (str,)
+    assert isinstance((int,) @ Ty((str,)), Ty)
+
+
+def test_probe_non_integer_domain():
+    from discopy.python.additive import Function, Ty
+
+    f = Function(lambda obj: obj.upper(), Ty((str,)), Ty((str,)))
+    assert Function.equation_factory(f, f)
+
+
+def test_axioms():
+    from discopy import testing
+    from discopy.python.additive import Function
+
+    testing.assert_axioms(Function)